+++ docs/run-dev.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-pip install -r requirements-dev.txt
-
-make livehtml