+++ test_classful/test_default_methods.py
@@ -1,6 +1,5 @@
 from flask import Flask
 from .view_classes import DefaultMethodsView, NoDefaultMethodsView
-from nose.tools import eq_
 
 app = Flask('default_methods')
 
@@ -11,14 +10,14 @@ NoDefaultMethodsView.register(app)
 def test_default_methods():
     client = app.test_client()
     resp = client.get('/default-methods/foo/')
-    eq_(b"GET", resp.data)
+    assert b"GET" == resp.data
     resp = client.post('/default-methods/foo/')
-    eq_(b"POST", resp.data)
+    assert b"POST" == resp.data
 
 
 def test_no_default_methods():
     client = app.test_client()
     resp = client.get('/no-default-methods/foo/')
-    eq_(b"GET", resp.data)
+    assert b"GET" == resp.data
     resp = client.post('/no-default-methods/foo/')
-    eq_(resp.status_code, 405)
+    assert resp.status_code == 405